refactor a bit to clean up render_main file
